// src/api/generated/models/Picture.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Picture = {
  id: number;
  entity: string;
  entity_id: number;
  is_main?: boolean;
  url: string;
  size?: number;
  updated_at: number;
  created_at: number;
};